+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added online help.
 - Add validation for duplicated boreholes during the cvs import.
 - Added functionality to add, edit and delete wateringresses, which belong to the hydrogeology observations of a borehole.
+- Use .NET API for managing borehole files.
 
 ### Changed